checkout: switch to pointer for all url.URL since gob seems to be buggy on value

+++ checkout/infrastructure/contextstore/redis_test.go
@@ -33,11 +33,11 @@ func getRedisStore(network, address string) *contextstore.Redis {
 	return new(contextstore.Redis).Inject(
 		new(flamingo.NullLogger),
 		&struct {
-			MaxIdle                 float64 `inject:"config:commerce.checkout.placeorder.contextstore.redis.maxIdle"`
-			IdleTimeoutMilliseconds float64 `inject:"config:commerce.checkout.placeorder.contextstore.redis.idleTimeoutMilliseconds"`
-			Network                 string  `inject:"config:commerce.checkout.placeorder.contextstore.redis.network"`
-			Address                 string  `inject:"config:commerce.checkout.placeorder.contextstore.redis.address"`
-			Database                float64 `inject:"config:commerce.checkout.placeorder.contextstore.redis.database"`
+			MaxIdle                 int    `inject:"config:commerce.checkout.placeorder.contextstore.redis.maxIdle"`
+			IdleTimeoutMilliseconds int    `inject:"config:commerce.checkout.placeorder.contextstore.redis.idleTimeoutMilliseconds"`
+			Network                 string `inject:"config:commerce.checkout.placeorder.contextstore.redis.network"`
+			Address                 string `inject:"config:commerce.checkout.placeorder.contextstore.redis.address"`
+			Database                int    `inject:"config:commerce.checkout.placeorder.contextstore.redis.database"`
 		}{MaxIdle: 3, IdleTimeoutMilliseconds: 240000, Network: network, Address: address, Database: 0})
 }
